refactoring type utils into its own module
